+++ src/ic-cdk-timers/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.7.0] - 2024-03-01
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.13.
+
 ## [0.6.0] - 2023-11-23
 
 ### Changed